chore(sidecar): move server and delegation implementation of the commitments-api in different folders
