+++ bolt-sidecar/src/api/commitments/delegation/mod.rs
@@ -1,0 +1,4 @@
+/// The processor of incoming commitments request via a websocket connection.
+pub mod processor;
+/// The Commitments receiver.
+pub mod receiver;
