+++ bolt-sidecar/src/api/commitments/mod.rs
@@ -1,14 +1,5 @@
-/// The commitments-API stream handler.
-pub mod firewall_recv;
-/// The commitments-API request handlers.
-mod handlers;
-/// The commitments-API headers and constants.
-mod headers;
-/// JSON-RPC helper types and functions.
-mod jsonrpc;
-/// The commitments-API middleware.
-mod middleware;
-/// The commitments-API JSON-RPC server implementation.
+/// The logic to accept commitments via firewall delegation, where a proposer specifies a list of
+/// RPC URLs from which it will accept commitments via a websocket connection.
+pub mod delegation;
+/// The Commitments-API server implementation.
 pub mod server;
-/// The commitments-API specification and errors.
-pub mod spec;
