+++ bolt-sidecar/src/api/commitments/server/middleware.rs
